+++ cmd/talosctl/cmd/mgmt/qemu_launch.go
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-//go:build linux
+//go:build linux || darwin
 
 package mgmt
